Refactoring all code, change struct of project files, remove useless code, optimizations and clearing output struct
