+++ internal/twoChannel/response_types.go
@@ -1,0 +1,19 @@
+package twoChannel
+
+type ResponsePosts struct {
+	Threads []struct {
+		Posts []struct {
+			Files []struct {
+				Name    string `json:"fullname"`
+				Path    string `json:"path"`
+				Preview string `json:"thumbnail"`
+			}
+		}
+	}
+}
+
+type ResponseThreads struct {
+	Threads []struct {
+		Id string `json:"num"`
+	}
+}
