--- sources/fourChannel/ResponsePosts.go
@@ -1,9 +1,0 @@
-package fourChannel
-
-type ResponsePosts struct {
-	Posts []struct {
-		Filename      string `json:"filename"`
-		FileExtension string `json:"ext"`
-		FileId        int64    `json:"tim"`
-	}
-}
