--- sources/fourChannel/ResponseThreads.go
@@ -1,7 +1,0 @@
-package fourChannel
-
-type ResponseThreads []struct {
-	Threads []struct {
-		Id int64 `json:"no"`
-	}
-}
