--- sources/twoChannel/ResponsePosts.go
@@ -1,13 +1,0 @@
-package twoChannel
-
-type ResponsePosts struct {
-	Threads []struct {
-		Posts []struct {
-			Files []struct {
-				Name    string `json:"fullname"`
-				Path    string `json:"path"`
-				Preview string `json:"thumbnail"`
-			}
-		}
-	}
-}
