--- sources/twoChannel/ResponseThreads.go
@@ -1,7 +1,0 @@
-package twoChannel
-
-type ResponseThreads struct {
-	Threads []struct {
-		Id string `json:"num"`
-	}
-}
