--- sources/types/Board.go
@@ -1,6 +1,0 @@
-package types
-
-type Board struct {
-	Name        string
-	Description string
-}
