--- sources/types/ChannelMessage.go
@@ -1,7 +1,0 @@
-package types
-
-type ChannelMessage struct {
-	VendorName string
-	Files      []File
-	Thread     Thread
-}
