--- sources/types/GrabberSchema.go
@@ -1,6 +1,0 @@
-package types
-
-type GrabberSchema struct {
-	Vendor Interface
-	Boards []Board
-}
