--- sources/types/Interface.go
@@ -1,7 +1,0 @@
-package types
-
-type Interface interface {
-	FetchThreads(board Board) (threads []Thread, err error)
-	FetchFiles(thread Thread) (files []File, err error)
-	VendorName() string
-}
