--- sources/types/Thread.go
@@ -1,14 +1,0 @@
-package types
-
-import (
-	"fmt"
-)
-
-type Thread struct {
-	ID    int64
-	Board Board
-}
-
-func (thread *Thread) StringId() string {
-	return fmt.Sprint(thread.ID)
-}
