+++ types/allowed_extensions.go
@@ -1,0 +1,16 @@
+package types
+
+type AllowedExtensions []string
+
+func (allowedExtensions AllowedExtensions) Contains(desiredExtension string) (result bool) {
+	result = false
+
+	for _, extension := range allowedExtensions {
+		if extension == desiredExtension {
+			result = true
+			break
+		}
+	}
+
+	return
+}
