+++ types/board.go
@@ -1,0 +1,6 @@
+package types
+
+type Board struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}
