+++ types/grabber_schema.go
@@ -1,0 +1,6 @@
+package types
+
+type GrabberSchema struct {
+	Vendor VendorInterface `json:"vendor"`
+	Boards []Board         `json:"boards"`
+}
