+++ types/thread.go
@@ -1,0 +1,6 @@
+package types
+
+type Thread struct {
+	ID    int64 `json:"id"`
+	Board Board `json:"board"`
+}
